cart 초안

+++ onlineshop/ecommerce/shop/migrations/0001_initial.py
@@ -54,7 +54,7 @@ class Migration(migrations.Migration):
                 ('slug', models.SlugField(max_length=100)),
                 ('description', models.TextField(blank=True)),
                 ('price', models.PositiveIntegerField()),
-                ('originalPrice', models.DecimalField(decimal_places=2, max_digits=10)),
+                ('originalPrice', models.PositiveIntegerField()),
                 ('available', models.BooleanField(default=True)),
                 ('stock', models.PositiveIntegerField()),
                 ('created_at', models.DateTimeField(auto_now_add=True)),
